+++ examples/t12_default_ports.rs
@@ -4,7 +4,7 @@
 //! [tutorial:](https://www.behaviortree.dev/docs/tutorial-advanced/tutorial_12_default_ports).
 //! [cpp-source:](https://github.com/BehaviorTree/BehaviorTree.CPP/blob/master/examples/t12_default_ports.cpp).
 
-use behaviortree::{EMPTY_STR, prelude::*};
+use behaviortree::prelude::*;
 use core::fmt::{Display, Formatter};
 use nanoserde::{DeJson, SerJson};
 
@@ -36,25 +36,18 @@ impl FromStr for Point2D {
 	type Err = BehaviorError;
 
 	fn from_str(value: &str) -> Result<Self, Self::Err> {
-		// remove redundant ' and &apos; from string
-		let s = value
-			.replace('\'', EMPTY_STR)
-			.trim()
-			.replace("&apos;", EMPTY_STR)
-			.trim()
-			.to_string();
 		// check for json marker
-		if let Some(stripped) = s.strip_prefix("json:") {
+		if let Some(stripped) = value.strip_prefix("json:") {
 			let res = DeJson::deserialize_json(stripped)?;
 			Ok(res)
 		} else
 		// check for json content
-		if let Ok(res) = DeJson::deserialize_json(&s) {
+		if let Ok(res) = DeJson::deserialize_json(value) {
 			Ok(res)
 		} else
 		// try conventional
 		{
-			let v: Vec<&str> = s.split(',').collect();
+			let v: Vec<&str> = value.split(',').collect();
 			let x = i32::from_str(v[0])?;
 			let y = i32::from_str(v[1])?;
 			Ok(Self { x, y })